Strict null checks @/contentScript/pageEditor (#8393)
* separate ensureContextMenu

* fix imports

* moves rest of exported backgroudn contextMenu functions to separate files

* fix tests

* overlay

* move remaining files

* move functions to separate files

* fix imports

* lint fixes

* separate ensureContextMenu

* fix imports

* moves rest of exported backgroudn contextMenu functions to separate files

* fix tests

* overlay

* move remaining files

* move functions to separate files

* fix imports

* lint fixes
