+++ src/contentScript/pageEditor/resetTab.ts
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2024 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { reactivateTab } from "@/contentScript/lifecycle";
+import { clearDynamicElements } from "@/contentScript/pageEditor/dynamic/clearDynamicElements";
+import { stopInspectingNativeHandler } from "@/contentScript/pageEditor/elementPicker";
+
+export async function resetTab(): Promise<void> {
+  stopInspectingNativeHandler();
+  await clearDynamicElements({});
+  await reactivateTab();
+}
